+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/datenlord/aligned-utils/compare/v0.2.0...HEAD
 
+## Changed
+
++ `AlignedBytes::into_raw` takes `this: Self`
+
 ## [0.2.0] - 2020-12-31
 
 [0.2.0]: https://github.com/datenlord/aligned-utils/compare/v0.1.5...v0.2.0